src/requirements.txt
asgiref==3.2.7
Babel==2.8.0
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
django-phone-verify==1.1.0
django-phonenumber-field==4.0.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
idna==2.9
nexmo==2.4.0
phonenumbers==8.12.1
PyJWT==1.7.1
python-dotenv==0.12.0
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
twilio==6.38.0
urllib3==1.25.8
gunicorn==20.0.4
psycopg2==2.8.5
dj-database-url==0.5.0
whitenoise==5.0.1
django-cleanup==4.0.0
